requirements.txt
asn1crypto==1.4.0
certifi==2021.10.8
cffi==1.13.2
charset-normalizer==2.0.8
cryptography==3.4.8
idna==3.3
Jinja2==2.11.3
MarkupSafe==1.1.1
oscrypto==1.2.1
pycparser==2.21
pycryptodomex==3.11.0
PyJWT==1.7.1
pyOpenSSL==19.1.0
pytz==2021.3
requests==2.26.0
six==1.16.0
snowflake-connector-python==2.7.1
urllib3==1.26.7
boto3
